Turn Pixel type param into an associated type
This does not compile at this point, more passes will be required to fix the
breakage, and I suspect that more traits can be converted to use associated
types afterwards: `ImageBuffer` for example, because most (all?) cases I have
seen now use it as `ImageBuffer<Container, P::Subpixel, P>`, which suggests
that `P::Subpixel` can actually be used directly now that it is an associated
type.

Furthermore, the `ConvertBuffer` impl for `ImageBuffer` cannot be expressed
correctly at this point, because equality constraints in where clauses are not
yet supported.