Refactor ImageBuffer to use less type params
This removes `into_dynamic`, which cannot be supported any more. It does reduce
the amount of type parameters, and it leads to a cleaner API in my opinion.
A cleanup pass is still needed, there might be repeated bounds. Also, a
`'static` bound on the pixel and subpixel types appears in many places, it
might need to be a constraint for the `Pixel::Subpixel` type? Furthermore, it
seems that different impls for the same struct, but with different type params
(some generic, some explicit) are not always recognised. Some specific code
could maybe be removed in favour of the generic impl.