Fix lifetime errors
I am unsure about whether `Pixel::Subpixel` and `GenericImage::Pixel` should
have a `'static` bound. It does not seem to cut help very much, as the `'static`
bound still needs to be specified in many places.

Also, many occurences of `GenericImage` now require a `'static` bound. I do not
know what the impact of this is for users of the library.