bmp/decoder: add full support for 16-bit bitfields
Allow any contiguous bitmasks, as well as alpha bitmasks.  Note that the
meaning of the alpha channel is not specified; it is not necessarily
transparency. It is up to the client to interpret it correctly.

As a result of this, we need to update some of the test reference
images to include the alpha channel.  Even though this channel is
not necessarily transparency, we still encode it as that in the png
reference images so that we can test the bmp decoder.