Add more checks to bmp decoder and limit the starting buffer size + minor optimization.
This commit adds some size limits to the decoder, limits the palette buffer size, and limits the starting pixel buffer size to avoid OOM issues.
Additionally, this adds a small optimisation to read_full_byte_pixel_data by reducing the number of read calls, which improves the performance of decoding images with plain 8-bit r/g/b/a values.
This commit also fixes the compilation of benchmarks which was broken.