Resize fixes (#661)
* Use absolute value of x in triangle filter

* Fix problems causing artifacts in image resizing

These artifacts are noticeable when using the Nearest filter to resize
images:
 - When resizing an image to its current size, the image becomes
   blurry.
 - When increasing the size of an image, it is also translated 1 pixel
   up and to the left, leaving black pixels along the right and bottom
   sides of the image.

The root of the problem was in the application of filters in resizing
operations. Specifically, it seemed like the calculation of the range of
input pixels affecting output pixels, and the numeric argument given to
filters, were incorrect.

This commit also adds comments to the resizing functions explaining the
calculations of these values.

* Remove artifacts of SIMD in image sampling

This removes the unnecessary vectorisation of sums and weights in image
sampling. It appears these vectors were once used in SIMD instructions,
but this is no longer the case.

* Don't scale up filter when downsampling

This fixes a problem when reducing an image's size where input pixels
were influencing the colour of output pixels further away than the
filter's supported window.