Fix size of sample when scaling.
The floating point "size" used for nearest should be 0, not 0.5.  Fix
clamping so at least one source pixel is used sampled for each output
pixel, but also so the floating point "size" applies to source size.

I'm not too sure of this for scaling UP, but for scaling _down_ it
seems more or less right.