Remove instances of unsafe code
This commit changes the interface of StreamingDecoder to not return references
to internal state. This makes it much easier to implement in a way that
satisfies the borrow checker, without resorting to unsafe.

It isn't completely clear whether this change is completely desirable because it
incurs additional copying overhead for some (uncommon?) use cases.