Extend mutable sample level access to `View`
This is a bit strange at first and might cause confusion which is why
the constructor method has such been chosen to have a long name. In
short, an image with mutable pixels requires packed samples while `View`
does not and this new `View` variant only gives access to mutable
samples.

Of course, this adds another extra converter for upgrading to `ViewMut`
at any later point. This is then slightly more efficient and fails only
to a still intact `View`.