Remove actual unsafety in hdr
Third part of the effort of removing `unsafe` from the library, the hdr
decoder actually required an interface change. Several other instances
had been removed in earlier commits. The method would use `Vec::set_len`
even though it did not have full control over the type parameter, as it
was a public interface.

Now simply requires the memory into which to emplace the result to be
provided and allocated externally. This makes the callers responsible
for correct optimization.

Removes the optimization for the library internal structures; it is
expected that the compiler picks up on the pattern well enough for it to
not be the main bottleneck.

Leaves the instances that are deprecated methods, they should enjoy some
deprecation period afterall and are not inherently bad as they are
correctly marked with `unsafe` themselves.

Co-Authored-By: Andreas Molzer <andreas.molzer@gmx.de>