Remove unecessary 'static bounds on affine operations
We only require the pixel itself to be static, probably as the
`ImageBuffer` will allocate it into a vector and never mention any
lifetime bounds in its interface? The complete removal of the static
bound is left to another set of patches since it requires modifying the
implementation of `ImageBuffer`.