Mark fractal example as no_run
A side effect of running the example in the readme is the creation of
the fractal.png output image. While we could also add this to gitignore,
the action of saving to `png` is also provided behind a (per default
enabled) feature flag. Marking it no_run avoids depending on feature
flags in the Readme.