Bump compiler requirements
The version supports edition=2018 and several `const` improvements. It
is also the one shipped with current Debian stable (Buster, released one
day ago).

The oldest version supporting the new edition is 18 weeks older than the
chosen one but the stabilization of TryFrom, several integer const
functions, integer atomics and NonZero_ types are not unlikely to get
used eventually. So let's bump it in one go, but no farther than Debian
stable.

**Note**: We will not always choose the minimum required version based
on Debian stable releases but they are considered an indicator on the
speed of update progression of systems outside the fast-paced Rust core
ecosystem. The minimum version should be regarded as a tracking
property, not a string guarantee. While we will try to keep changes out
of point release, there is no such guarantee and security considerations
have priority.