Provide docker files for building the fuzzer
Ensures that the build is locally reproducible. Libfuzzer and fuzzit
integration relies on a set of supported versions. At the current moment
gcc 9.0 is too _new_ for example leading to breakage with an up-to-date
ArchLinux environment. Creation of new fuzzing binaries should/will be
eventually performed on CI but it's nice to have a local solution
available.