Fix compatibility with 'long-term' stable version
Usage of `copied` is unecessary in the context of appending a slice to a
vector. While potentially a replacement it relies on many internal
details to guarantee the same performance. Where it is used as the first
initialization of a Vec it is definitely nicer to convert the source
into an owned version and even moreso when the source is a Vec itself.