Convert error with private methods, not From
The latter is part of the a publicly visible crate interface. This makes
it impossible to upgrade the image format crates used for these specific
format over major versions. Using this pattern is problematic in several
aspects:

* less mature formats are harder to adopt as they are a huge commitment
* bugs and issues that require breaking changes require a coordinate
  effort across the whole dependency chain and ecosystem
* dependent crates we control do not evolve as independently as I would
  wish

It should arguably be avoided to have an unecessary public interface
dependency *unless a format reaches a stable 1.0 interface itself* where
it is no large liability. For consistency this should be uniform across
formats without regards to whether they are internally implemented, a
crate in image-rs, or otherwise.