Revert breaking interface change
One clippy suggestion in 55b1b285a8d57039a64296ba5ef66a9658a0e21d was a
breaking change. I think the expectation was that the next release were
a major one. But an encoder bug should probably be fixed. The risk of
breakage for a small API change as this seems low but why take chances?