Update the quickcheck dev-dependency
It has become quite outdated at this point but luckily is maintained
with a commitment to MSRV of 1.30. The main point of the update are some
reported issues with rand and the dependency rdrand. Not that any of
those are actually relevant but this reduces the burden of proof for
verifications (crev, etc.) of code bases using the image crate.