Do a documentation pass and improve naming
All `new` constructors accept a detailed underlying description while
the shorthands and alternatives are renamed to `from_*_hint` and
`from_kind` respectively. This is supposed to encourage utilizing a
detailed error representation in dependencies with the new constructor.

Note that the UnsupportedError and LimitError have no `new` constructor
but utilize the naming convention regardless. This is to avoid blocking
the name if we introduce an underlying error representation for them.