Add a saturating conversion from time::Duration
Given the prior usage, it does not make much sense to provide an exact
conversion. However, it should be clear that precision is not finite and
the representation is not exactly that of a Duration.

The implementation turned out to be fairly involved as the num-fraction
crate does not provide approximation algorithms for fractions. No fear,
we code them ourselves with extensive testing to cover much more than
the required accuracy tests.