Add an owning stream buffer
To avoid duplication or breaking the interface the internal
representation is an enum that may either hold a reference with lifetime
or owns the buffer with `'static` lifetime. This allows adding a new
interface that should be as good as a truly owning the writer, down to
the lifetime and even `Send` and `Sync` bounds (as contrary to Cow there
is no shared reference involved).