Rewrite unfiltering for speed
The code structure was previously unfit for automatic vectorization.
Some experiments on Godbolt shows that llvm can now do the required
transformation, or at least unroll some loops.

The problematic part were indices: It was hidden that the length would
always be a multiple of `bpp`. Furthermore, each loop iteration appeared
dependent on the immediate predecessor by reading from the array we've
been writing. This is precisely what mutable reference in Rust are
supposed to improve. They are a proof to the optimizer that certain
statements can be executed independently and concurrently.

Since this mutably borrows the slice we can not trivially access the
previous bytes for predictors. However, move semantics enable us to
store each chunk after we're done so that the next can then access it in
the next loop iteration. Quite neat.

Additionally, the method was not optimized according to the fixed
possibilities for bpp there are. Using macros we now monomorphize and
optimize the different variants independently.