Lazily cleanup row buffer
This delays moving the remaining buffer of raw scan lines, until the
buffer space must be reused for the next chunk of decoded data. It then
performs a single `memmove` instead.

The commit also changes initialization of the initial previous row to
occur with a single call to `Vec::resize` instead of pushing zeroes
which uses memset instead of a loop.