Fix prediction requiring unfiltered previous line
All filter functions are specified in terms of Orig, which is defined:
> Orig(y) denotes the orginal (unfiltered) value of byte y

The previous row must thus be remember the unfiltered values instead of
being set to the written filtered bytes. Instead of moving the filtered
bytes after compressing we keep a slice referring the previous raw of
input data, which is correct as there is no intermediate transformation
applied on the library level.