Address lint of missing Copy
This lint (missing_copy_implementations) no longer seems to trigger in
more recent version of Rust. Some issues [#46695][1] seem to indicate
that this is based on detection of externally visible types that might
have improved—note that the types in question are private to their
module.

It should not be any risk to add this impl as it is not externally
visible. Consider the scenario where the types are made public by
extracting the modules into their own crate. In this case, the
discussion on the exhaustiveness of these representations will need to
occur in any case and be more SemVer restrictive than the Copy impl.

[1]: (https://github.com/rust-lang/rust/issues/46695)