Fix decoding apng chunks before requested
The decoder iterates over the scan lines and reads them into the
provided buffer. In the non-interlaced case it would rely on the image
itself to end after all scan lines are encoded. However, the method to
supply next scanlines would skip over chunk boundaries. This meant that
when an image was requested we would also decode all following APNG
chunks (and ignore their contents). This would also fully decode and
check them, with wrong expectation of the contents.

We now stop after having decoded the whole image, i.e. remember the
height and iterate only over those scanlines.

This also fixes another inconsistency that prevented interlaced decoding
to detect a checksum error since it terminated the decoding loop earlier
than the standard decoding and did not witness the end event of the
chunk.