Implement the fixed iterator
Just like the `Pixels` iterator we do not yield any rows when the rows
of the image are empty (even if height is not 0), as otherwise we would
have to ensure that `height` can be represented as a `usize` to satisfy
the requirements of `ExactSizeIterator`. (Not literally, we could always
panic or return an incorrect results without causing unsoundness.
However that would be a suboptimal solution when we can already be
consistent with prior art).