Fix reading fdAT sequence number only once
This would have previously been read at the start of every single
partial chunk that was to be passed into the inflater, i.e. with each
pass of data from the underlying buffered stream. Instead we want to
read a sequence number exactly once at the start of the chunk.