Correctly flush&stop image data on IDAT end
All data in an image must appear in consecutive IDAT/fdAT chunks that
are not interrupted by any other data chunk. In particular we should not
ignore intermediate chunks but report them as an error (see 5.6. Chunk
Ordering). This is also important to properly flush the zlib
decompression at the end of the subframe and retrieve remaining rows as
well as check its checksum even if ignored blocks follow.