Make apng infallible to convert
This fits the model of e.g. emulsion better: We can query in advance if
the image is animated and can then decide to decode it as an animation
but get the pure PNG fallback if we are oblivious to the animation
capabilities just as if `png` itself did not support it. An ImageDecoder
adapter can also be added as a rather pure extra.

It's also cleaner to shift errors to actual decoding of images. We can
delay allocation of buffers until that point. The exact buffer
requirements could also depend on the dynamic information gained for the
next image which would make it hard to query limits at construction time
in any case.