Check buffer sizes for overflow
Fixes an issue affecting mainly platforms with 32-bit (or smaller)
pointer size where the calculation of required buffer sizes would
use unchecked math ops that could silently overflow.

We now try to calculate the maximum of the necessary lengths up-front,
as part of converting to a Reader, which ensures that all other
computations have no overflow.