Provide scale method of JpegDecoder.
Add a `scale(&mut self, requested_width, requested_height)` to
`JpegDecoder`, exposing same functionality in `jpeg_decoder::Decoder`.

Question: As it is allowed for this function to do nothing, maybe it
should be added to trait `ImageDecoder`?  If would be possibe to add a
meaningfull implementation to any "progressive" format, such as png
and gif, and it's ok to do nothing for other formats.