Add ImageOutputFormat support for AVIF
PR #1314 added encoding support for the AVIF image format. The PR however did not yet include ImageOutputFormat support. An ImageOutputFormat entry is required when encoding and writing images through DynamicImage::write_to. This commit adds an Avif entry to the ImageOutputFormat enum as well as a conversion from an ImageFormat to an ImageOutputFormat.