Split write_segment into write_segment and write_marker
This allows us to get rid of data: Option<&[u8]> in write_segment, which makes a big difference for the overall code size and removes the necessary checks for whether data is Some() or None.

This additionally removes the need to use byteorder::{BigEndian, WriteBytesExt}, by using to_be_bytes instead.