Turn NeuQuant into newtype around color_quant type
This ensures coherence in downstream crates and keeps the SemVer
promises of the interface. Otherwise, we could have the following
situation:

```
// In some downstream:
trait ColorMap {}
impl ColorMap for color_quant::NeuQuant {}
impl ColorMap for image::math::nq::NeuQuant {}
```

This would be an incompatible impl if our type was merely an alias for
the NeuQuant type from `color_quant`.