Add Adaptive filtering method for encoding
Implement Adaptive filtering which uses the heuristic described in
the PNG standard. For each line, all five filter algorithms are executed
on the buffer. Filtered bytes are then summed by the absolute value of
their i8 representation. The filter with the lowest sum is selected as
the "best" filter for that row of pixels.

An Option<FilterType> of None signifies the Adaptive encoding filter.

Change the Info struct filter field to an Option<FilterType>
Change encoder::filter signature to take an Option<FilterType>
encoder::filter now returns a FilterType