Fix trusting TGA data on data length
When doing Run-Length encoding a packet might describe more data than we
expect in the decoded image. Since we copy this to the user output
buffer we must however decode the exact amount of pixel data, otherwise
such a copy will panic due to out-of-bounds.

We truncate the decoded data to the expected size. Note that RLE packets
should not cross scanlines so truncation should also work for decoding
of a single scanline.