Remove feature png-encoding
After switching to miniz_oxide for encoding there will no longer be any
advantage in this feature, thus we may as well remove it. This split
only reduces the dependency tree by a minuscule amount anyways.

The main goal here is to reduce the interface. In particular there was
already one new major release that we couldn't track without one of our
own. The loss is that one can no longer precisely configure the level of
compression which might be remediated with a new setter.