Accept any AsRef<[u8]> for PLTE, tRNS and ICC data
Currently, a user of this crate, when encoding a PNG file with custom
PLTE, tRNS or ICC data, is forced to pass that data as an owned Vec<u8>.
However, the internals of this library do not really need an owned
vector, just something that can be converted to a byte slice, and
the requirement of a Vec may introduce unneeded copies and allocations
in client code.

Fix the situation by adding type parameters that allow accepting any
type that implements the AsRef<[u8]> trait.