Address StreamWriter concerns introducing a PartialInfo struct
Now Writer only receives a struct with a subset of the data contained in
the Info struct. This data excludes the palette and the transparency
chunks, as they are only needed once, when calling init(), where an
actual Info struct reference may be passed.

As the new PartialInfo struct does not contain any lifetime parameters,
Writer doesn't need any lifetime parameters and the logic of associated
structs is greatly simplified, allowing all of the StreamWriter methods
to work again.