Work on meta data API consistency
The fields of Metagram are now public, since it is not intended that
they do keep invariants relative to one another. The documentation also
clarifies the relation to the actual processing applied by the decoding.
The SharedRecorder now has the same basic set of methods for adding meta
data as the original Recorder.
Added documentation and some examples.