Return OutputInfo per image
This fixes a long standing sore point of the interface where the caller
needed to remember a tuple returned from read_info, instead of just the
required buffer size, while the actual frame information would still
vary frame by frame, which had to be retrieved manually.

We still require a large enough buffer to hold any frame even when a
single frame does not use this full buffer. This is to make the
interface more robust. Otherwise a call to `next_frame` is more at risk
of panic due to a buffer too small as there is no way to know
beforehand. We could also enable public visibility for `init` which
decodes all information necessary for the byte size of the next frame
but I'm not yet too sure about being more permissive just because of a
(potentially unrelated) extra method call.