Add check on tga::decoder to prevent crash
Previously the `pixel_data` buffer was assumed to be the same length as
the parameter buffer passed to `ImageDecoder::read_image`. This PR puts
a check in place to make sure that both buffers are the same length
before attempting to copy one into the other, this prevents triggering a
panic when calling `slice::copy_from_slice`.