Simplify/Rework logic of frame tracking
We no longer _enforce_ that the number of frames is exactly as deduced
by animation control/separate default image. This isn't necessary for
a correct chunk stream. It's not exactly correct chunk ordering though:

>  	Multiple IDAT chunks shall be consecutive

Instead, we leniently allow encoding of multiple images. When we run out
of animation frames we simply switch back to IDAT. The behavior when no
animation frames had been added should then be the one of version 0.16.