Fix breakage from mp4parse update
This was a breaking release published in a minor version. We used two
APIs that had breakage: AvifDecoder no longer has a public `alpha_item`
and we now interpret the length of the alpha data instead (following the
documentation of the C-API) and `read_avif` takes a parameter on how to
handle decoding failure where we used the normal mode for now until we
get indicated to be strict or lenient.