Remove 'static bound from traits and impls
The 'from traits' part is the breaking change. Downstream consumers
could have previously made use of this additional bound.

This is not required, we never utilize types `&'static [Subpixel]` or
similar and if we would then it could be a local bound. But the bound is
infectious in that it requires placing several additional bounds in
every place where the `Pixel` trait is being used.