Fix image-rs#1618: bmp::decoder::extend_buffer didn't properly handle certain buffer sizes.
The function had a working strategy for cases where the buffer was being doubled or more,
but if the buffer wasn't being doubled, the part of the function that copied the old data to the
end of the buffer absolutely blew up.

Along with a simpler implementation that should always work, there's a regression test case so
this specific area doesn't break again. Thanks @saethlin!