Add TINFLStatus message
The `TINFLStatus` value was not used which caused a warning when
compiling. I have added a message for each status and a catch-all in
case a new one is added. That way the user will have a better idea of
what caused the deflate stream to be corrupt.

I tested that the Adler32 checksum message was shown by opening the
image from issue #294 with the show example app.