Succeed early on unusual resize quickcheck case
The quickcheck now exits early when checking resizing when it runs into
the case where we are limited by `u32::MAX`. These cases should still be
handled fine but in general should probably be covered by a separate
test. In particular we can no longer expect the new size to be exact in
the requested dimension.