Fix transparency error handling
The original code created an error but never returned it. Also, it
didn't use `OutsidePlteIdat` error variant so this also fixes an unused
variant warning.

The code now checks that the transparency chunk is in the right place
and returns an error message if it isn't. The transparency chunk is
supposed to be after a PLTE chunk and before an IDAT chunk according
to the [specification](https://www.w3.org/TR/PNG/#5ChunkOrdering).