Move color/bit depths combination check
There was a TODO comment in the `parse_ihdr` method to check if the
color/bit depths combination is valid. It turned out there was already
a check for this in the `read_info` method. It makes more sense for all
of the checks to be in one place, so I moved the check into the
`parse_ihdr` method.

I also renamed the fields of the `InvalidColorBitDepth` error enum
value to match their names in the rest of the code. That way the field
names do not need to be named when constructing the
`InvalidColorBitDepth` value. The type in not exposed publicly so it
is not an API breaking change.