Assign instead of `replace`
There is no need to use `replace` or `swap` to swap the buffers when the
old value is no longer needed. It can just be assigned instead. It makes
more sense and there is no return value that needs to be ignored. Thanks
to @fintelia for pointing out that `swap` was not needed either.