Add duplicate chunk check
The duplicate chunk error variant was unused and resulted in a warning
when compiling. Only some chunk types are allowed to appear multiple
times according to the [specification](https://www.w3.org/TR/PNG/#5ChunkOrdering).
I have added checks for that. This also fixes the unused variant warning
when compiling.

The `parse_chunk` method is already checking if the info is `None` so it
is safe to unwrap the info in the subsequent `parse_*` methods.

Ideally the parsing should be refactored at some point so that the
parsing methods only receive what they need and not a reference to the
whole mutable self. For example, the header, which creates the info, must
appear right after the signature. That means that the info could be
passed to the subsequent `parse_*` methods and the they would not need
to unwrap it. That would, however, be a large complicated change.