Compensate for overfull buffer in write
Output formats, but mainly PNG, may expect the byte buffer to be
precisely sized according to the image descriptor. The `ImageBuffer`
implementation permits its underlying container to be larger than this
for multiple reasons, and won't check this property in `from_raw`.
Hence, the actual size may be too large. This mismatch can cause
encoding to fail for a curious reason:

	failed to save: IoError(Custom { kind: Other, error: "wrong data size, expected 1288872 got 1289808" })

Too much data is harmful?

We avoid this by pre-selecting only the relevant portion of the
underlying container.