cargo-deny
[cargo-deny](https://github.com/EmbarkStudios/cargo-deny) is an amazing tool that protects from:

* duplicated crates (code bloat)
* copy-left licenses in the dependency tree
* RUSTSEC advisories

In this case it discovered the duplication of 'jpeg-decoder' and 'miniz_oxide', both because of the 'tiff' crate.

Until we can update the tiff crate and resolve those problem I opted to ignore these duplicated crated in the deny.toml file.

Adding cargo-deny to the CI will prevent more dependency duplication, hence this unsolicited PR.