Use SWAR in Sub decoding
The idea for SWAR is to treat regular integer types like `u32|u64` as a SIMD
register for arbitrary smaller-width bit lanes. We can calculate 3 bytes per
pixel and 4bpp pixels within a `u32`, 6bpp and 8bpp pixels within a `u64`.

It's not an n-times speedup based on the number of lanes because SWAR may
require some overhead of maintaining the lanes depending on the operation
(shifts, for instance). This is evidenced by masking operations during
calculation to correct or prevent lanes from overlapping. For the wrapping
addition operation here, 7-bit arithmetic is calculated to prevent a carry
digit affecting neighboring lanes. After the addition, a correcting XOR addition
is done to repair the most significant bit and arrive at the correct answer for
8-bit wrapping unsigned arithmetic.