Fix to prevent blocking file::open that may happen on some file types.
File::open can hang when opening a pipe file since it blocks until there is some data to read. If the pipe is empty or not ready, the program will wait indefinitely. To prevent this we can pass custom flags to OpenOptions which is what File::Open is an abstaction of. The changes allow for optionally setting custom flags when opening a file using : image::open_with_custom_flags() image::image_dimensions_with_custom_flags(), reader::open_with_custom_flags() functions that take a path and customflag as arguments. This has been done in a way so as to not introduce any braking changes.