Independent `row_stride_in_bytes` in `expand_adam7_bits`.
This commit refactors implementation of `expand_adam7_bits` so that it
can accept `row_stride_in_bytes` that is different from the expanded
width of the frame or image.

This is helpful for making `expand_pass` work in scenarios where
interlaced row needs to expanded into a bigger image (e.g. if the
currently decoded frame is an animation frame that only takes a
subregion of the whole image - in this case the stride between expanded
rows is bigger than the size of expanded rows).

And the above is helpful for exposing `expand_pass` through a public API
that will hopefully pass the test of time...